HS-244 | Update for languageString refactor
