+++ core/src/main/java/pl/agh/edu/ui/component/label/LanguageLabel.java
@@ -7,7 +7,7 @@ import pl.agh.edu.ui.language.LanguageManager;
 import pl.agh.edu.utils.LanguageString;
 
 public class LanguageLabel extends CustomLabel implements LanguageChangeListener {
-	private final LanguageString languageString;
+	private LanguageString languageString;
 
 	public LanguageLabel(LanguageString languageString, String font) {
 		super(font);
@@ -16,8 +16,8 @@ public class LanguageLabel extends CustomLabel implements LanguageChangeListener
 		onLanguageChange();
 	}
 
-	public void updateLanguagePath(String languagePath) {
-		this.languagePath = languagePath;
+	public void updateLanguagePath(LanguageString languageString) {
+		this.languageString = languageString;
 		onLanguageChange();
 	}
